+++ js/repl/fable-library-js/AsyncBuilder.d.ts
@@ -39,21 +39,21 @@ export interface IAsyncContext<T> {
     cancelToken: CancellationToken;
     trampoline: Trampoline;
 }
-export type IAsync<T> = (x: IAsyncContext<T>) => void;
-export declare function protectedCont<T>(f: IAsync<T>): (ctx: IAsyncContext<T>) => void;
-export declare function protectedBind<T, U>(computation: IAsync<T>, binder: (x: T) => IAsync<U>): (ctx: IAsyncContext<U>) => void;
+export type Async<T> = (x: IAsyncContext<T>) => void;
+export declare function protectedCont<T>(f: Async<T>): (ctx: IAsyncContext<T>) => void;
+export declare function protectedBind<T, U>(computation: Async<T>, binder: (x: T) => Async<U>): (ctx: IAsyncContext<U>) => void;
 export declare function protectedReturn<T>(value: T): (ctx: IAsyncContext<T>) => void;
 export declare class AsyncBuilder {
-    Bind<T, U>(computation: IAsync<T>, binder: (x: T) => IAsync<U>): (ctx: IAsyncContext<U>) => void;
-    Combine<T>(computation1: IAsync<void>, computation2: IAsync<T>): (ctx: IAsyncContext<T>) => void;
-    Delay<T>(generator: () => IAsync<T>): (ctx: IAsyncContext<T>) => void;
-    For<T>(sequence: Iterable<T>, body: (x: T) => IAsync<void>): IAsync<void>;
-    Return<T>(value?: T): (ctx: IAsyncContext<T | undefined>) => void;
-    ReturnFrom<T>(computation: IAsync<T>): IAsync<T>;
-    TryFinally<T>(computation: IAsync<T>, compensation: () => void): (ctx: IAsyncContext<T>) => void;
-    TryWith<T>(computation: IAsync<T>, catchHandler: (e: any) => IAsync<T>): (ctx: IAsyncContext<T>) => void;
-    Using<T extends IDisposable, U>(resource: T, binder: (x: T) => IAsync<U>): (ctx: IAsyncContext<U>) => void;
-    While(guard: () => boolean, computation: IAsync<void>): IAsync<void>;
+    Bind<T, U>(computation: Async<T>, binder: (x: T) => Async<U>): (ctx: IAsyncContext<U>) => void;
+    Combine<T>(computation1: Async<void>, computation2: Async<T>): (ctx: IAsyncContext<T>) => void;
+    Delay<T>(generator: () => Async<T>): (ctx: IAsyncContext<T>) => void;
+    For<T>(sequence: Iterable<T>, body: (x: T) => Async<void>): Async<void>;
+    Return<T>(value: T): (ctx: IAsyncContext<T>) => void;
+    ReturnFrom<T>(computation: Async<T>): Async<T>;
+    TryFinally<T>(computation: Async<T>, compensation: () => void): (ctx: IAsyncContext<T>) => void;
+    TryWith<T>(computation: Async<T>, catchHandler: (e: any) => Async<T>): (ctx: IAsyncContext<T>) => void;
+    Using<T extends IDisposable, U>(resource: T, binder: (x: T) => Async<U>): (ctx: IAsyncContext<U>) => void;
+    While(guard: () => boolean, computation: Async<void>): Async<void>;
     Zero(): (ctx: IAsyncContext<void>) => void;
 }
 export declare const singleton: AsyncBuilder;